refactor: move isArtsyEmail to utils
